+++ examples/simulate_hybrid_wbattery_dispatch.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from hopp.simulation.sites import SiteInfo, flatirons_site
 from hopp.simulation import HybridSimulation
-from hopp.dispatch.plot_tools import plot_battery_output, plot_battery_dispatch_error, plot_generation_profile
+from hopp.simulation.technologies.dispatch.plot_tools import plot_battery_output, plot_battery_dispatch_error, plot_generation_profile
 from hopp.utils.keys import set_nrel_key_dot_env
 # Set API key
 set_nrel_key_dot_env()
