+++ hopp/simulation/technologies/__init__.py
@@ -1,9 +1,9 @@
+from .power_source import PowerSource
 from .battery import Battery
 from .clustering import Clustering
 from .csp_plant import CspPlant
 from .detailed_pv_plant import DetailedPVPlant
 from .grid import Grid
-from .power_source import PowerSource
 from .pv_plant import PVPlant
 from .tower_plant import TowerPlant
 from .trough_plant import TroughPlant
