+++ hopp/simulation/technologies/dispatch/plot_tools.py
