+++ hopp/simulation/technologies/dispatch/power_sources/pv_dispatch.py
@@ -5,7 +5,7 @@ import PySAM.Pvsamv1 as Pvsam
 import PySAM.Pvwattsv8 as Pvwatts
 import PySAM.Singleowner as Singleowner
 
-from hopp.dispatch.power_sources.power_source_dispatch import PowerSourceDispatch
+from hopp.simulation.technologies.dispatch.power_sources import PowerSourceDispatch
 
 
 class PvDispatch(PowerSourceDispatch):
