+++ hopp/simulation/technologies/dispatch/power_sources/trough_dispatch.py
@@ -3,7 +3,7 @@ import datetime
 
 import PySAM.Singleowner as Singleowner
 
-from hopp.dispatch.power_sources.csp_dispatch import CspDispatch
+from hopp.simulation.technologies.dispatch.power_sources import CspDispatch
 
 
 class TroughDispatch(CspDispatch):
