+++ hopp/simulation/technologies/dispatch/power_storage/power_storage_dispatch.py
@@ -2,7 +2,7 @@ import pyomo.environ as pyomo
 from pyomo.network import Port
 from pyomo.environ import units as u
 
-from hopp.dispatch.dispatch import Dispatch
+from hopp.simulation.technologies.dispatch import Dispatch
 
 
 class PowerStorageDispatch(Dispatch):
