+++ hopp/simulation/technologies/wind_plant.py
@@ -3,9 +3,11 @@ import PySAM.Windpower as Windpower
 import PySAM.Singleowner as Singleowner
 from hopp.add_custom_modules.custom_wind_floris import Floris
 
-from hopp.simulation.technologies.power_source import *
+from hopp.simulation.technologies.power_source import PowerSource
+from hopp.simulation.sites import SiteInfo
 from hopp.simulation.layout.wind_layout import WindLayout, WindBoundaryGridParameters
-from hopp.dispatch.power_sources.wind_dispatch import WindDispatch
+from hopp.simulation.technologies.dispatch import WindDispatch
+from hopp.utils.log import hybrid_logger as logger
 
 
 class WindPlant(PowerSource):
